Move event_loop conftest to the existing tests module conftest
