+++ cylc/uiserver/tests/conftest.py
@@ -14,11 +14,11 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Test code and fixtures."""
 
+import asyncio
 import inspect
 
 import pytest
 import zmq
-
 from cylc.flow.network import ZMQSocketBase
 
 
@@ -43,3 +43,16 @@ class AsyncClientFixture(ZMQSocketBase):
 @pytest.fixture
 def async_client():
     return AsyncClientFixture()
+
+
+@pytest.fixture
+def event_loop():
+    """This fixture defines the event loop used for each test."""
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    # gracefully exit async generators
+    loop.run_until_complete(loop.shutdown_asyncgens())
+    # cancel any tasks still running in this event loop
+    for task in asyncio.all_tasks(loop):
+        task.cancel()
+    loop.close()
